update code style
