+++ include/rmoss_mindvision_driver/mindvision_cam_node.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RM_CAM__MINDVISION_CAM_NODE_HPP
-#define RM_CAM__MINDVISION_CAM_NODE_HPP
+#ifndef RMOSS_MINDVISION_DRIVER__MINDVISION_CAM_NODE_HPP_
+#define RMOSS_MINDVISION_DRIVER__MINDVISION_CAM_NODE_HPP_
 
 #include <string>
 #include <memory>
@@ -23,25 +23,25 @@
 #include "rmoss_cam/cam_server.hpp"
 #include "rmoss_mindvision_driver/mindvision_cam.hpp"
 
-namespace rmoss_entity_cam
+namespace rmoss_mindvision_driver
 {
-    // Node warpper for MindVisionCamera
-    class MindVisionCamNode
-    {
-    public:
-        explicit MindVisionCamNode(
-            const rclcpp::NodeOptions &options = rclcpp::NodeOptions());
+// Node warpper for MindVisionCamera
+class MindVisionCamNode
+{
+public:
+  explicit MindVisionCamNode(
+    const rclcpp::NodeOptions & options = rclcpp::NodeOptions());
 
-        rclcpp::node_interfaces::NodeBaseInterface::SharedPtr get_node_base_interface()
-        {
-            return node_->get_node_base_interface();
-        }
+  rclcpp::node_interfaces::NodeBaseInterface::SharedPtr get_node_base_interface()
+  {
+    return node_->get_node_base_interface();
+  }
 
-    private:
-        rclcpp::Node::SharedPtr node_;
-        std::shared_ptr<rmoss_entity_cam::MindVisionCam> cam_dev_;
-        std::shared_ptr<rmoss_cam::CamServer> cam_server_;
-    };
-} // namespace rm_cam
+private:
+  rclcpp::Node::SharedPtr node_;
+  std::shared_ptr<rmoss_mindvision_driver::MindVisionCam> cam_dev_;
+  std::shared_ptr<rmoss_cam::CamServer> cam_server_;
+};
+}  // namespace rmoss_mindvision_driver
 
-#endif // RM_CAM__MINDVISION_CAM_NODE_HPP
+#endif  // RMOSS_MINDVISION_DRIVER__MINDVISION_CAM_NODE_HPP_
