+++ launch/mindvision_cam.launch.py
@@ -1,16 +1,17 @@
-#  Copyright 2022 RoboMaster-OSS
+#!/usr/bin/python3
+# Copyright 2022 RoboMaster-OSS
 #
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-#      http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import os
 
@@ -18,14 +19,17 @@ from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
 from launch_ros.actions import Node
 
+
 def generate_launch_description():
 
-    param_path = os.path.join(get_package_share_directory("rmoss_mindvision_driver"), "config/cam_param.yaml")
+    param_path = os.path.join(
+        get_package_share_directory('rmoss_mindvision_driver'),
+        'config/cam_param.yaml')
 
     mv_cam_node = Node(
-        package="rmoss_mindvision_driver",
-        executable="mindvision_cam",
-        name="mindvision_camera",
+        package='rmoss_mindvision_driver',
+        executable='mindvision_cam',
+        name='mindvision_camera',
         parameters=[param_path],
         output='screen'
     )
