--- src/mindvision_cam_node_main.cpp
@@ -1,28 +1,0 @@
-// Copyright 2022 robomaster-oss.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include <memory>
-
-#include "rclcpp/rclcpp.hpp"
-#include "rmoss_mindvision_driver/mindvision_cam_node.hpp"
-
-int main(int argc, char **argv)
-{
-    rclcpp::init(argc, argv);
-    auto node = std::make_shared<rmoss_entity_cam::MindVisionCamNode>();
-    rclcpp::spin(node->get_node_base_interface());
-    rclcpp::shutdown();
-
-    return 0;
-}
